+++ tomcat/src/main/java/nextstep/jwp/controller/Controller.java
@@ -2,8 +2,8 @@ package nextstep.jwp.controller;
 
 
 import org.apache.coyote.http11.request.RequestMethod;
-import org.apache.coyote.http11.request.mapping.RequestMapping;
-import org.apache.coyote.http11.request.mapping.RequestParam;
+import org.apache.coyote.http11.request.annotation.RequestMapping;
+import org.apache.coyote.http11.request.annotation.RequestParam;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpStatus;
 import org.apache.coyote.http11.response.Resource;
